Fix merge issue and move BlockChainInfo to better place
